// src/types/ui.ts
export type UnoColorVariants =
  | 'rose'
  | 'pink'
  | 'fuchsia'
  | 'purple'
  | 'violet'
  | 'indigo'
  | 'blue'
  | 'sky'
  | 'cyan'
  | 'teal'
  | 'emerald'
  | 'green'
  | 'lime'
  | 'yellow'
  | 'amber'
  | 'orange'
  | 'red'
  | 'gray'
  | 'slate'
  | 'zinc'
  | 'neutral'
  | 'stone'

export type UISizeVariants = 'xs' | 'sm' | 'md' | 'lg' | 'xl'
export type UIColorVariants =
  | 'default'
  | 'primary'
  | 'secondary'
  | 'third'
  | 'fourth'
export type UIRoundedVariants =
  | 'none'
  | 'xs'
  | 'sm'
  | 'md'
  | 'lg'
  | 'xl'
  | 'full'